backend/app/requirements.txt
addict==2.4.0
fastapi==0.115.12
fiftyone==1.4.1
gradio==5.25.0
huggingface_hub==0.29.1
ipdb==0.13.13
matplotlib==3.10.1
numpy==2.2.3
Pillow==11.2.1
psutil==7.0.0
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
setuptools==77.0.3
supervision==0.25.1
torch==2.5.1
torchvision==0.20.1
termcolor==3.0.1
timm==1.0.15
tqdm==4.67.1
transformers==4.30.2
typer==0.15.2
yapf==0.43.0
